--- abstractions/IEditExerciseMetadata.ts
@@ -1,5 +1,0 @@
-interface EditExerciseMetadata {
-    exerciseName: string;
-    workout_date: number;
-    logged_exercise_id: number;
-}
