--- abstractions/IExerciseLog.ts
@@ -1,8 +1,0 @@
-export interface ExerciseLog {
-    weight_log_id: number;
-    weight_logged: string; 
-    reps_logged: string;
-    set_number: number; 
-    workout_date: number; 
-    day_name: string 
-}
